Refactor MCP into separate server file
